[fix] 2단계 피드백 반영 -price, quantity Wrapper 클래스 적용

+++ src/test/java/kitchenpos/product/domain/ProductFixture.java
@@ -6,7 +6,7 @@ import java.math.BigDecimal;
 
 public class ProductFixture {
 
-    public static Product 상품(Long id, String name, BigDecimal price) {
+    public static Product 상품(Long id, String name, int price) {
         Product product = new Product(name, price);
         ReflectionTestUtils.setField(product, "id", id);
         return product;
